requirements.txt
flask>=3.0.0
google-genai>=1.0.0
yfinance>=0.2.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
